fix: resolve 4 frontend issues
- Restore Product Detail Fetching: add safe price arithmetic with Number() conversion and total price display
- Fix Error Handler Module Split: extract pure error mapping to src/lib/blockchainError.ts, keep re-exports for backward compatibility
- Align API Base URL And Endpoint Contract: use centralized API_BASE_URL in logger, correct API.md paths to match server routes
- Make Checkout Token Contract Configuration Consistent: consolidate all token lookups into networkConfig helpers, remove duplicate currencyToTokenContract

Closes #530
Closes #531
Closes #532
Closes #533
